+++ c7n/c7ncli/group/integrations.py
@@ -2,6 +2,7 @@ import click
 
 from c7ncli.group.integrations_dojo import dojo
 from c7ncli.group.integrations_sre import sre
+from c7ncli.group.integrations_chronicle import chronicle
 
 
 @click.group(name='integrations')
@@ -11,3 +12,4 @@ def integrations():
 
 integrations.add_command(dojo)
 integrations.add_command(sre)
+integrations.add_command(chronicle)