+++ c7n/c7ncli/group/ruleset_eventdriven.py
@@ -12,19 +12,17 @@ def eventdriven():
 
 
 @eventdriven.command(cls=ViewCommand, name='add')
-@click.option('--version', '-v', type=float, default=1.0,
-              help='Ruleset version')
 @click.option('--cloud', '-c', type=click.Choice(RULE_CLOUDS),
               required=True, help='Ruleset cloud')
-@click.option('--rule_id', '-rid', multiple=True, required=False,
-              help='Rule ids to attach to the ruleset')
-@click.option('--rule_version', '-rv', type=str, required=False,
-              help='Rule version to choose in case of duplication '
-                   '(the highest version by default). Used with '
-                   '--full_cloud or --standard flags')
+@click.option('--version', '-v', type=str,
+              help='Ruleset version to create. If not specified, will be '
+                   'resolved automatically based on the previous ruleset '
+                   'with the same name or based on rulesource of type '
+                   'GITHUB_RELEASE')
+@click.option('--rule_source_id', '-rsid', required=False, type=str,
+              help='The id of rule source object to use rules from')
 @cli_response()
-def add(ctx: ContextObj, version: float, cloud: str,
-        rule_id: tuple, rule_version: str, customer_id):
+def add(ctx: ContextObj, cloud, version, rule_source_id, customer_id):
     """
     Creates Event-driven ruleset with all the rules
     """
@@ -32,8 +30,8 @@ def add(ctx: ContextObj, version: float, cloud: str,
     return ctx['api_client'].ed_ruleset_add(
         version=version,
         cloud=cloud,
-        rule=list(rule_id),
-        rule_version=rule_version
+        rule_source_id=rule_source_id,
+        customer_id=customer_id
     )
 
 
@@ -58,8 +56,9 @@ def describe(ctx: ContextObj, cloud: str, get_rules: bool, customer_id):
 @click.option('--cloud', '-c', required=True,
               type=click.Choice(RULE_CLOUDS),
               help='Event-driven ruleset cloud to describe')
-@click.option('--version', '-v', type=float, default=1.0,
-              help='Event-driven ruleset version to delete')
+@click.option('--version', '-v', type=str, required=True,
+              help='Event-driven ruleset version to delete. '
+                   '\'*\' to delete all versions')
 @cli_response()
 def delete(ctx: ContextObj, cloud: str, version: float, customer_id):
     """